Add runtime exception for playwright archive not existing

+++ lib/live_load/browser/connection/playwright/decompressor.ex
@@ -16,6 +16,15 @@ defmodule LiveLoad.Browser.Connection.Playwright.Decompressor do
 
   defp do_extract(version) when is_binary(version) do
     archive = Application.app_dir(:live_load, ["priv", "playwright", version, "playwright_bundle.tar.gz"])
+
+    if not File.exists?(archive) do
+      raise RuntimeError, """
+      The Playwright bundle does not currently exist.
+
+      In order to install Playwright, run `mix live_load.install` to install the default version.
+      """
+    end
+
     dest = Application.app_dir(:live_load, ["priv", "playwright", version, "bin"])
     File.rm_rf!(dest)
     File.mkdir_p!(dest)
